pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "bumblebee"
version = "0.1.0"
description = "Research codebase for post-transformer architecture exploration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3,<2.0",
    "omegaconf>=2.3,<3.0",
    "torch>=2.2",
    "tensorboard>=2.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["adaptive_inference*", "hierarchical_programs*", "memory_state*", "shared*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]
